Searches successfully log relevant results
