+++ src/frontend/package.json
@@ -6,9 +6,13 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.24.0",
     "buffer": "^6.0.3",
+    "cors": "^2.8.5",
+    "express": "^4.18.1",
     "googleapis": "^101.0.0",
     "https": "^1.0.0",
+    "nodemon": "^2.0.16",
     "prettier": "^2.6.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
@@ -18,6 +22,7 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
+    "start:backend": "nodemon searchResults.js",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
